use react-component
